Formatting files since it was getting cluttered
